src/bot/commands/task_commands.py
"""Task management Discord commands."""

from typing import Any

import discord
import structlog
from discord.ext import commands

logger = structlog.get_logger(__name__)


class TaskCommands(commands.Cog):
    """Task management commands for Discord bot."""

    def __init__(self, bot: Any):
        self.bot = bot
        self.logger = logger.bind(component="TaskCommands")

    @commands.command(name="task_help")
    async def task_help(self, ctx: commands.Context) -> None:
        """Show task management help."""
        embed = discord.Embed(
            title="📋 Task Management Commands",
            description="Available task commands",
            color=discord.Color.blue(),
        )

        embed.add_field(
            name="!task_create",
            value="Create a new task",
            inline=False,
        )
        embed.add_field(
            name="!task_list",
            value="List active tasks",
            inline=False,
        )
        embed.add_field(
            name="!task_complete",
            value="Mark task as completed",
            inline=False,
        )

        await ctx.send(embed=embed)

    @commands.command(name="task_create")
    async def task_create(
        self, ctx: commands.Context, *, task_description: str
    ) -> None:
        """Create a new task."""
        try:
            # Task creation logic would go here
            embed = discord.Embed(
                title="✅ Task Created",
                description=f"Task created: {task_description}",
                color=discord.Color.green(),
            )
            await ctx.send(embed=embed)
            self.logger.info("Task created", task=task_description)
        except Exception as e:
            self.logger.error("Failed to create task", error=str(e))
            await ctx.send("❌ Failed to create task")

    @commands.command(name="task_list")
    async def task_list(self, ctx: commands.Context) -> None:
        """List active tasks."""
        try:
            embed = discord.Embed(
                title="📋 Active Tasks",
                description="No active tasks found",
                color=discord.Color.blue(),
            )
            await ctx.send(embed=embed)
        except Exception as e:
            self.logger.error("Failed to list tasks", error=str(e))
            await ctx.send("❌ Failed to list tasks")

    @commands.command(name="task_complete")
    async def task_complete(self, ctx: commands.Context, task_id: str) -> None:
        """Mark task as completed."""
        try:
            embed = discord.Embed(
                title="✅ Task Completed",
                description=f"Task {task_id} marked as completed",
                color=discord.Color.green(),
            )
            await ctx.send(embed=embed)
            self.logger.info("Task completed", task_id=task_id)
        except Exception as e:
            self.logger.error("Failed to complete task", error=str(e))
            await ctx.send("❌ Failed to complete task")
