pyproject.toml
[project]
name = "mindbridge"
version = "0.1.0"
description = "MindBridge: Personal AI-powered knowledge management system for Discord and Obsidian integration"
authors = []
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"

dependencies = [
    "discord.py>=2.5.2",
    "google-genai>=0.2.0",
    "google-cloud-speech>=2.21.0",
    "google-cloud-secret-manager>=2.20.0",
    "garminconnect>=0.2.28",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.0",
    "asyncio-throttle>=1.0.2",
    "pyyaml>=6.0.2",
    "rich>=13.9.0",
    "structlog>=24.4.0",
    "tenacity>=9.1.2",
    "pydub>=0.25.1",
    "python-dateutil>=2.9.0.post0",
    "numpy>=2.1.0",
    "scikit-learn>=1.5.0",
    "beautifulsoup4>=4.12.3",
    "requests>=2.32.0",
    "cryptography>=45.0.7",
    "google-api-python-client>=2.181.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "I"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.mypy]
python_version = "3.13"
strict = false
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
    "pytest-cov>=6.0.0",
    "ruff>=0.12.8",
    "mypy>=1.17.1",
    "types-python-dateutil",
    "types-pyyaml",
    "types-aiofiles",
    "types-requests",
    "types-beautifulsoup4",
    "pre-commit>=4.3.0",
]
